backend/requirements.txt
boto3
flask
flask-cors
pandas
uwsgi

python-dotenv